gauntlet: add a cheat to test ST version in arcade
In the ST version of gauntlet, players start with 2000 HP and not 600!
This makes the game playable alone, at least the 1st levels.
Notice there was a forgotten else in init_cpuid which made me wonder for
a long time why cpu_id was initialized with CPU_6502!
The commit also includes a fix which was not needed after all for
set_cpu_id, because ram should be initialized too when changing cpu_id
in the console.
The cheat is for the warrior only for now, ideally it should be replaced
by some rom hack to change the default HP setting for all players, but
this was done quickly... !