more precision when reporting a watch...
when watching an address on more than 2 bytes to look for a long value
change, for example $9048d6 in gauntlet to check the HP of the warrior,
the value is in starscream longs, meaning that the 1st word is the most
significant one, 0 most of the time. So if you want a meaningful result
when reporting the value change, you must display the long at this
address!
Notice this code is for 68000 only, if ever I want to do that for the
68020 or sh2, I'll have to change the ReadLongSc here...